% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegHammersmith-package.R
\docType{package}
\name{ggsegHammersmith-package}
\alias{ggsegHammersmith}
\alias{ggsegHammersmith-package}
\title{ggsegHammersmith: Hammersmith Brain Atlas for the 'ggseg' Ecosystem}
\description{
Hammersmith maximum probability map with 83 regions for the 'ggseg' ecosystem. Provides a unified 'ggseg_atlas' object with both 2D polygon geometry and 3D vertex indices, for use with 'ggseg' and 'ggseg3d'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegHammersmith}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegHammersmith/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

}
\keyword{internal}
