% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hammersmith}
\alias{hammersmith}
\title{Hammersmith Brain Atlas}
\usage{
hammersmith()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Maximum probability map with 83 regions from 30 healthy adult subjects.
}
\examples{
hammersmith()
plot(hammersmith())
}
\references{
Hammers A, et al. (2003). Three-dimensional maximum probability
atlas of the human brain. \emph{NeuroImage}, 19(4):1301-1315.
\doi{10.1016/S1053-8119(03)00165-7}
}
\concept{ggseg_atlases}
