% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hammersmith_cortical}
\alias{hammersmith_cortical}
\title{Hammersmith Cortical Atlas}
\usage{
hammersmith_cortical()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Cortical regions from the maximum probability map with 83 regions
from 30 healthy adult subjects.
}
\examples{
hammersmith_cortical()
}
\references{
Hammers A, et al. (2003). Three-dimensional maximum probability
atlas of the human brain. \emph{NeuroImage}, 19(4):1301-1315.
\doi{10.1016/S1053-8119(03)00165-7}
}
\seealso{
Other ggseg_atlases: 
\code{\link{hammersmith_subcortical}()}
}
\concept{cortical_atlases}
\concept{ggseg_atlases}
